Fail on test 17, add phony target to Makefile
all but test 17 passed. it requires sanitizer to debug, but
`make SANITIZER=1` makes it not passing test 15,
when sorting an 10^6-element array, the process unexpectedly aborts
without any error message, reason unknown

+++ queue.c
@@ -148,7 +148,7 @@ bool q_remove_head(queue_t *q, char *sp, size_t bufsize)
  * Return number of elements in queue.
  * Return 0 if q is NULL or empty
  */
-size_t q_size(queue_t *q)
+int q_size(queue_t *q)
 {
     if (q == NULL || q->head == NULL)
         return 0;
